+++ client/src/components/EpicEntryItem/index.tsx
@@ -1,11 +1,8 @@
 import React, { useState } from 'react';
 import S from './style';
 import draggableIcon from '@public/icons/draggable.svg';
-import { Modal } from '@/lib/design';
 import { EpicType } from '@/types/epic';
 import EpicEditModal from '../EpicEditModal';
-import { useEpicDispatch } from '@/lib/hooks/useContextHooks';
-import { updateEpicById } from '@/lib/api/epic';
 
 interface EpicEntryItemProps {
   activated: boolean;
@@ -22,21 +19,8 @@ const EpicEntryItem = (props: EpicEntryItemProps) => {
   const [showEditModal, setShowModal] = useState(false);
   const [value, setValue] = useState(props.epicData.name);
 
-  const dispatchEpic = useEpicDispatch();
-
   const handleChange = (ev: React.ChangeEvent) => {
     setValue((ev.target as HTMLInputElement).value);
-  };
-
-  const handleFormSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    const updatedEpic = { ...props.epicData, name: value };
-    updateEpicById(props.epicData.id, updatedEpic);
-    dispatchEpic({
-      type: 'UPDATE_EPIC',
-      epic: updatedEpic,
-    });
-    setShowModal(false);
   };
 
   return (
@@ -51,20 +35,13 @@ const EpicEntryItem = (props: EpicEntryItemProps) => {
         <S.DragIndicator src={draggableIcon} alt="draggableicon" showDraggable={showDraggable} />
         <div>{props.epicData.name}</div>
       </S.Container>
-      <Modal
-        shouldConfirm
-        title="에픽 정보 수정"
-        visible={showEditModal}
-        onClose={() => setShowModal(false)}
-        onClickOk={handleFormSubmit}
-      >
-        <EpicEditModal
-          epicData={props.epicData}
-          value={value}
-          handleChange={handleChange}
-          handleFormSubmit={handleFormSubmit}
-        />
-      </Modal>
+      <EpicEditModal
+        showEditModal={showEditModal}
+        setShowModal={setShowModal}
+        epicData={props.epicData}
+        value={value}
+        handleChange={handleChange}
+      />
     </>
   );
 };
